+++ packages/ts-newskit/src/slices/types.ts
@@ -1,0 +1,12 @@
+export type SliceArticle = {
+  id?: string;
+  headline: string;
+};
+
+export type MouseEventType = React.MouseEvent<HTMLAnchorElement, MouseEvent>;
+
+export type ClickHandlerType = (
+  event: MouseEventType,
+  article: SliceArticle,
+  position?: string
+) => void;