+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.15.85](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.15.84...@times-components/author-profile@6.15.85) (2023-08-10)
+
+**Note:** Version bump only for package @times-components/author-profile
+
+
+
+
+
 ## [6.15.84](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.15.83...@times-components/author-profile@6.15.84) (2023-08-08)
 
 **Note:** Version bump only for package @times-components/author-profile